Removed the population of the Model routes from the constructor and provided a public AddValidRoute() method to add them, to make the network more extensible.
This required the removal of the Port enum.  A list of valid ports is now stored in a HashSet and used for validation.

Defined the ports as constants in the unit tests.

Improved the naming of the unit tests to make them clearer.

Added unit tests that check for exceptions.

Migrated three separate unit tests into a single CalculateShortestJourney_Tests() with three [TestCase] attributes.